requirements.txt
###### Requirements without Version Specifiers ######`
selenium
webdriver-manager
fastapi
uvicorn[standard]
python-dotenv